supervisr/mod/auth/oauth/client/requirements.txt
django-all-access==0.9.0
pycrypto>=2.4
requests_oauthlib>=0.4.2
oauthlib>=0.6.2
